Bugfix to prevent invalid interactive state
The issue is that the state of the page may start out in 'no_interactive' but with the show_interactive checkbox checked. When the box is unchecked, it tries to send a 'remove_interactive' event to a state which doesn't have one. The solution is to use the "back" button to return to the edit page and check the box again; this sends an "add_interactive" event, which generates a new interactive and moves the state to "has_interactive". From there, we can un-check the box and send the "remove_interactive" event to the right state.
[#39101151]