Different trigger for automatic modal editing.
Instead of automatically launching the modal edit window(s) when the page is loaded and certain fields are blank, this launches the modal edit windows only when certain GET parameters are in the request, e.g. "?edit_embed_mc=22" will automatically launch the edit for the multiple-choice embeddable with ID 22. This parameter is added to the URL after a new Embeddable or Interactive is added to the page, and it is stripped from subsequent requests. (In most cases, the parameter is simply not carried over into the next request; in a few cases, however, where the controller action would redirect "back" to the previous URL, we're actually stripping all GET parameters from that redirect to avoid re-launching the edit modal.) This is for [#40312235].