Correctly re-sizes interactives [#49075313]
The "native" height and width set in the edit mode are used to calculate an aspect ratio which is included with the iframe tag as a data attribute. The width of the iframe is set to 100%, so it fills its container. There are a few lines of JS which run on page load which determine the aspect ratio and the actual width, and adjust the height accordingly.