Fixes interactive scrolling into the footer
The behavior described in [#49854249] was partly due to the un-pinning of the interactive being fixed to the top of the related-content div. On many pages, that div is never rendered, so the interactive never reached the bottom of its track. Adding a dummy div to mark the bottom of the scrolling track, which is always in the DOM, fixes the issue.

https://www.pivotaltracker.com/story/show/49854249