WIP: Refactoring embeddables carousel JS
To finish [#49877927], this pulls all JS for handling the carousel out into its own module for testability. Configuration of the carousel now starts with the controls disabled and enables them selectively depending on where we are in the list (i.e. if we're at list start, "previous" is disabled; at list end, "next" is disabled). Methods are in place for adjusting the size of the carousel, but not fleshed out. Spec tests are outlined but incomplete (and currently not passing).

https://www.pivotaltracker.com/story/show/49877927