Adjust height and width of embeddables carousel for layout
[#49877927] This adjusts the height and width of the carousel window and its component embeddables according to the page layout. 

Height is set to the height of the interactive (if the layout puts the carousel next to the embeddable) or the shorter of the tallest embeddable or the available space (if the layout puts the carousel under the window). The embeddable height is adjusted to leave room for the buttons underneath. 

The width is set according to the layout: 960px (full-width), 374 (60-40), or 278 (70-30).

https://www.pivotaltracker.com/story/show/49877927