HTML for Sequence views [#51592849]
This gets all the content on the page. In the case of the index page, where the style is borrowed from the Activities listing, it's pretty close to "finished". The new and edit forms aren't bad but don't account for adding and removing activities (yet). The actual show page puts all the content on but is close to unstyled.

https://www.pivotaltracker.com/story/show/51592849