Tweaks the flow of sequence creation [#52539193]
Associating activities with a sequence is the most important part of creating them, but we can't do that association on the "new" page because the sequence doesn't yet exist on the server. This follows a pattern I used for page creation, creating an "empty" sequence with default values and redirecting to the edit view for that sequence rather than spending any time on the "new" view.

https://www.pivotaltracker.com/story/show/52539193