Use @project and @theme for header, footer and CSS
Note that this changes the landscape for [#51730923] because the link in the upper left is now that of the project, and the logo is as well. It's still the case that it's unclear how to get to the LARA index page, but I'm not convinced we need to get to the LARA index page from the runtime.

[#52741509] [#52056969]

https://www.pivotaltracker.com/story/show/52741509
https://www.pivotaltracker.com/story/show/52056969
https://www.pivotaltracker.com/story/show/51730923