Activities are sequence-aware
If an activity is launched from a Sequence page, the sequence id is part of the URL. If the sequence ID is part of the URL, the sequence is stored as part of the Run data, and whenever the user comes back to that run, the activity sets the appropriate @activity variable for use by whatever needs it (e.g. theming).

[#52741275]

https://www.pivotaltracker.com/story/show/52741275