"Save" and "Edit" buttons look different in IE
This fixes [#54311176]. The jeditable editor produced buttons with type "submit" and "cancel" for save and cancel, respectively, but IE munges the "cancel" type into "submit". Because I was using the type as a CSS selector, this meant both link/buttons got the same CSS and thus the same position. Here I've fed jeditable HTML to use for its buttons which is similar but also carries class attributes which should be untouched by IE and make it possible to distinguish the two in CSS.

https://www.pivotaltracker.com/story/show/54311176