Updates behavior of sequence navigation menu
Another step in [#51593657]. This makes all the activity links in the menu change the "active" activity, i.e. the one with the blue highlighting and the visible list of pages. Final step should be to get this behavior with the carets/arrows and the previous behavior with the text links.

https://www.pivotaltracker.com/story/show/51593657