Don't add interactive-pinning waypoints if not needed
For [#53568273] and may also help with [#53579129]. This checks to see if there's actually enough height in the page for the interactive to move if it's pinned, or if it would wind up running down over footer content (or back up over header content). If there's not enough space, the waypoints aren't set. This should fix a lot of weird scrolling behavior when the interactive block is close to the height of the info/assessment block (or taller than it).

https://www.pivotaltracker.com/story/show/53568273
https://www.pivotaltracker.com/story/show/53579129