Clear the save indicator if there are no questions
This fixes [#58915442]. The save indicator object in JS, which controls the state of the indicator on the page, is cleared on creation, but if no questions were found, it's never created and thus the indicator on the page remains in its initial state (suggesting nothing will be saved, which is technically correct but needlessly alarming). This creates the JS object if it hasn't been already, solely to clear the indicator.

https://www.pivotaltracker.com/story/show/58915442