Adds "drawing prompt" to ImageQuestion
for [#58968130]. This adds the `drawing_prompt` field to ImageQuestions, exposes it in the ImageQuestion form for editing, and shows it in the authoring view and the activity view.

While doing this, I removed all buttons but "Done" from the question dialog (this addresses [#58992462] but we need to add a "cancel" button to finish it), and also showed the background image in the runtime view if there isn't already an annotated image (i.e. if the user hasn't given an answer yet) which covers [#59778098].

https://www.pivotaltracker.com/story/show/58968130
https://www.pivotaltracker.com/story/show/58992462
https://www.pivotaltracker.com/story/show/59778098