WIP on ImageQuestion buttons [#58992462].
Simplifies the "Edit answer" button to just "Edit", and creates a "Replace snapshot" button for existing snapshot questions. Also adds a "cancel" button to the dialog which reverts changes to the text box if there is one, and while it doesn't roll back changes to the SVG it doesn't save them either. The two catches are that ("Cancel" doesn't set the SVG back to the saved state) and that if there's no text box for an image question, they may not see the "Edit" button even if they have provided an image answer. (It's hard to tell when a question has been answered in a way which is universal across the types of ImageQuestion.)

https://www.pivotaltracker.com/story/show/58992462