Failing spec test for Embeddable::Answer module.
I was looking at this, because there seemed to be an attempt to memoize
the result of prompt_no_itals into class vars.

This isn’t currently happening, but in addition to that, this prompt_no_initials
isn’t doing what we thought anyway.

One idea is to just remove references to @@cleaned_prompt?? TBD