Pinned interactives that work! [#59668812]
Instead of trying to handle images differently from other Interactive types, this solution simply does the naive setup of objects and the waypoints as though the Interactive is at its final size, but then it sets up a timer and runs, at intervals (every 2.5 seconds at the moment) a script which checks to see if the height of the Interactive has changed. If it has, all waypoints are cleared and re-set. (Because setting the waypoints checks if pinning is still appropriate, they won't be set if the Interactive is now too tall.)

I would prefer to listen for an event (e.g. onresize) generated by the changing size of the Interactive div, but there isn't such an event for div elements in Javascript at the moment.

https://www.pivotaltracker.com/story/show/59668812