Allow logins to choose the portal to auth with.
This should be the last piece of [#55550912]. If a user goes to the sign-in page ( /users/sign_in/ ) the login form and registration links are now hidden, and instead there's a row of links to configured portals to authorize with. (The register/sign-in forms are still available in development.) These links each have a parameter which designates which portal to use for authentication. (Actually the parameter designates an entry in a hash of allowable portals.) The Omniauth strategy then authenticates with the correct portal.

https://www.pivotaltracker.com/story/show/55550912