Reads portal configuration from YAML [#55550912].
This version is flat, to deliver just the name => url mapping. The next step will be to provide more data per configuration. Note that the keys are now strings, not symbols, throughout. (Also note that when we have more data, this YAML file will get moved into a "sample" and the real one will be deployed via Chef.)

https://www.pivotaltracker.com/story/show/55550912