Confirming page positions after activity copy.
This is an attempted fix for [#62930230]. Because `InteractivePage#set_list_position` saves without validating, it seems to also skip any `acts_as_list` efforts to maintain the integrity of the list. This adds a method `LightweightActivity#fix_page_positions` which sets the `position` value for each page according to its index in the `pages` array (which seems circular, because that array is supposed to be ordered according to the position value, but on a new Activity that can't be trusted).

Worth noting: that `fix_page_positions` method can also be used in the console to patch up activities which somehow arrive at the state described in [#62463830].

https://www.pivotaltracker.com/story/show/62930230
https://www.pivotaltracker.com/story/show/62463830