ActivityController & SequenceController #show goes to latest work.
By default the show actions in these controllers will bring the user to
wherever they last left off in sequences and activities.

These actions used to show the TOC (table of contents) aka index for the
activity or sequence.

The TOC is now only displayed if the param [:show_index] is passed in.

Part of [#55931526 ]

https://www.pivotaltracker.com/story/show/55931526