LARA user can see who they are logged in as  [#83759888].
 - A LARA user can clearly see who they are logged in as when they are
running an activity.
Hi, <username> Clicking the username will open a Pop up link that says
“Not you? Click here” where “here” leaves this window, signs you out
of LARA and the portal, and goes to the portal to sign in - same portal
that the last user came from.

 - For anonymous users will be named anonymous
Hi, Anonymous Pop up link that explains about creating an account
“If you are taking a class, check in with your teacher about logging in
to save data.”

 - Change the CONTROLLER_ACTION in auth_portal.rb to store the username
in the session as [:portal_username] obtained from the portal.

 - add DOM for showing the username in _activity_nav.html.haml.

 - add after_sign_out_path_for method to handle the user signout.
If re_login param is true then sign_out request is sent to the portal
which current user belong to.

 - handle re_login parameter in home_controller.rb and its rspec
If re_login param is true a request is sent to the portal obtained from
athe param 'provider'.