Add option for "preflighting" an iframe interactive
Certain browsers (IE and Safari) won't load cookies for a third-party
domain within in iframe unless the user has been to that domain in a
normal browser window. Pre-flighting pops up a window with a specific
url under that domain which initializes cookies for that domain and then
closes itself. Once that is complete, the iframe can be loaded and
cookies will work as expected.

[#96215356]
[#92480624]