use tagged rails logger for delayed job messages
there didn’t seem to be an ideal solution here, this approach lets you
see the AR log messages and our own log messages. And this approach also
lets us use the same block of code in a Job as well as normally.
An alternative approach is to try send all Job messages to delayed_job.log
however in that case the Rails AR logger needs to be overridden.