move the unauthorized code into runs namespace
the action to handle the unauthorized feedback form was in the pages
controller. However that route required an activity and page id to
generate the correct URL. Additionally the code was used by both pages
and now activities. By moving the code to the runs controller the route
is more simple, and now the code is in a shared place to be used by both
pages and activities.

Additionally this fixes an issue with the theme not being set correctly
when the unauthorized template was rendered by from the single_page
action.