simplify the question_trackers_controller
this does put all of the hash serialization of the question_tracker
into the controller. However, perviously some of it was in reporter
and some of it was the controller. 

If all of it should be outside of the controller then the trackers_json 
could be moved back to the reporter, and the controllers url_options can
be passed in. The report can use Rails.application.routes.url_helpers to
access the url helper in the class method.