remove now broken list of activities from info
This is an admin only method that is rarely used. It was broken by a
recent change that removed `question_answers`. It could work with just
`answers`. But that would require modifying the QuestionStandin to support
the `#activity` method. That did not seem worth doing for this rarely
used method.