Added interactiveStateUrl iframe-saver initInteractive message [#128441585]
This is needed so that the v2 launch page can add the url to the codap launch url.

Also fixed the get/put url helpers for interactive run state api endpoints.  When the endpoints were moved until the /api/v1 namespace the url helpers needed to have the namespace prepended to them.