Added user idle check in authoring pages
This adds a new /api/v1/user_check endpoint that returns either null if the user is not logged in or the user object if they are logged in.  A modified third party library was used to perform the idle checks.  When the user comes back from being idle the user_check endpoint is called and if it fails then a blocking modal message is shown with a _blank targeted link to re-login.  Once the user logs back in and come back to the page the user check is automatically called which should then dismiss the modal if they are logged in.