Fix interactive printing
Remove the extra div with a image in the background, this doesn’t work when printing.
Fix the img-container layout so it is visible when the iframe is not.
Add margins so the printed interactives line up better with the questions.
Fix ITSI printing so the interactives don’t shift around in relation to questions, and they are centered.
Fix some typos in comments about page layout labels.
Make note of issue with bold text when using the print_blank path.
Add a Node script to generate 24 page activity to check various layouts and interactive settings
Slightly widening the questions in full width mode.
[#142314353]