Add two optional overrides for connecting with the portal
The -proxy override is the more simple of the two. The user just needs to define the PORTAL_HOST env variable.
The -net override requires the user to find the the docker gateway ip,
add a host entry to their local /etc/hosts,
and add environment variables for the gateway, host, and port.