fix failing tests
These failures were related to the simplification of the reportable logic.
A interactive is only reportable if it has enable_learner_state.
If it just has_report_url it is not reportable.
The javascript code that asks the interactive for its state is only enabled if enable_learner_state
is true. And this state is where the report_url is stored, this is why enable_learner_state is
required.

An interactive could use the portal REST api to save its state instead
of the iframe postMessage. In this case it is possible for the interactive to save state
even if there enable_learner_state is disabled. But I don't think we need to support that case in
a special way. It is simpler to require authors to check the 'enable_learner_state' box in order to
get some form of reporting.

[#150113706]
[#150113578]