Fix another failing test
LARA sends two different types of hashes to the portal.
The recent changes to the logic about when to send each type, broke this test.
LARA now doesn't just look at the interactive state, to determine the type. Instead it
looks at the has_report_url flag on the MwInteractive model.

[#150113706]
[#150113578]