capitalization error was causing inconsistent authoredState
All of the interactive that I could find were already handling both strings and object
forms of the authoredState. So this fix should not break anything.

During runtime the authoredState was being sent to the interactive as a object
even with this bug. This is because the jquery data method is returning an
object not a string at least in Chrome. I'm not sure if this is a feature of jquery or of
Chrome. The server is sending down `data-authored-state` attribute, but in the
DOM this attribute is not present. Instead there is a dataset property on the element.

[#155754828]