[DL/NP] STASH: Vey WIP: Capybara tests for `mode=teacher` params.
Start of Initial LARA Teacher Edition:

Add a new mode to LARA for launching a teacher edition of an activity and sequence. Initially this edition will just include a banner saying 'Teacher Edition'. This banner needs to show on every single page of the activity. This includes sequence and activity home pages.

[#160836352]
https://www.pivotaltracker.com/story/show/160836352