log the amount of time a request sits in the queue
this should be a good metric for measuring the load on the server.
This metric is the time the request is waiting from when nginx receives it to when rack picks it up.
There could be additional queue time by the AWS load balancer, but that shouldn't be necessary
to get a good sense of how far behind the rack is getting.

[#162059682]