[DL/NP] : Fix: external-scripts block plugin editing.
🐞 FIX for this exception:

```
  An ActionView::Template::Error occurred in embeddable_plugins#edit:
  undefined method `is_hidden?' for #<Embeddable::ExternalScript:0x00000004573178>
```

Triggered when an author tries to edit a plugin on a page with an `Embeddable::ExternalScript`

These are the old style plugins such as the model feedback plugin.

They implemented `is_hidden` instead of `is_hidden?`.

New plugin edit pages need to find a list of all embeddables in the page that aren't hidden.  It sends `is_hidden?` to each embeddable.
This caused the above exception.

The fix was to add the method `is_hidden?` to `Embeddable::ExternalScript`

[#164970025]

https://www.pivotaltracker.com/story/show/164970025