Added plugin authoring metadata [#166193534]
Adds authoring_metadata to approved scripts which currently has list of components scoped to activity, embeddable and embeddable-decoration.  The authoring_metadata field is a stringified JSON blob to allow for change during the remaining plugin authoring stories.  Once the json is finalized a schema validator should be added as part of the code that loads the authoring metadata from the external json manifest.

The activity forms that deal with plugins now select plugins appropriate to the scope and then saves the component label to the plugin so that the component name can be used and displayed instead of the generic plugin name.  This allows for plugins to expose multiple components at different scopes.

One final change is that once a component is added at either the activity or embeddable level there is no way to change the type of component in the edit form.