Moved logger init to before plugin init [#166135835]
Moved the logger.js initialization to before the plugin initialization so that the plugins an use the logger.

Added an event queue to the logger so that events generated before the plugin pub/sub system is setup are not lost.

Note: before the logger was waiting for document.ready to initialize itself to ensure the window.gon variable existed.  Since the logger is only used in runtime.js and the runtime.js file is loaded after window.gon is set there is no need to wait for document.ready.