Plugin editing sets modal's css class in applicationController
This allows controls which use `respond_with_edit_form` (which the
embeddablePluginsController was doing via the embeddableController) to
pass down additional css classes to be used in the modal dialog.

The `allow-full-width` class removes the set width, letting the plugins
render to their desired width. This also improves the modal dialog
centering.