Add inline authoring to activity-level plugins
This uses the `plugins/_author` view for inline editing of >v3 plugins,
just like the embeddable plugins use. The author view has been modified
slightly to allow for the lack of `@embeddable`.

If the plugin is <v3, it will continue to use the form.

The plugins controller has been update to redirect back to the activity's
edit page after the form submits, instead of trying to render the plugin's
own page.