Use `answered?` method to check if answer should be sent to report service
We don't want to send all answers. Just the answers that the students have actually worked on.

Our previous approach was to compare updated_at and created_at timestamps. That wasn't working though, in part because of how sequence runs work.

Instead we will use answers answered? method. It should work for most things, with the possible exception of labbook answers.

[#166769565]

https://www.pivotaltracker.com/story/show/166769565