Attempt to reduce cognitive complexity of the perform function.
Code climate gave me a 9 on cognitive complexity for `perform` here.

Not sure if its the pre-emptive returns, or the `unless` syntax.  The docs aren't super clear
on what can be done in situations like this.  I started to break the method up, but it started
to feel more complex to me.

https://docs.codeclimate.com/docs/cognitive-complexity

[#165204332]

https://www.pivotaltracker.com/story/show/165204332