before_create was running before the embeddable_plugin had an id
Without an id, then the setup of the embeddable.plugin.plugin_scope was failing.
This only happened during duplicate. I didn't track down exactly why, but I'd guess
in other cases the embeddable was being saved first so it actually had an id.
Perhaps this happens on the embeddable.plugin = Plugin.create({}) call.

Switching to after_create makes sure the embeddable has an id. However
as the comment here says it'd be better to remove the duplication here.

Part of this code should also include a migration to fix the existing plugins that don't
have a valid plugin_scope