add new REPORT_SERVICE_TOOL_ID env variable
this allows a developer working locally to override the toolId and sourceKey that is sent to the report service
this way when multiple developers are all working with a local lara named app.lara.docker their data
in the report service does not conflict.

the tests have not been updated for this.
I'd also like to see if I can remove the REPORT_SERVICE_SELF_URL option so there is less for developers
to configure.