add number and bulleted lists to toolbar
This also adds aligncenter to the small editor toolbar.

The lists plugin was not added to tinymce, so the items weren't showing up,
even though they were configured.

In the fullsize editor this change caused the toolbar to have two lines, so I
removed strikethrough which isn't used much and alignjustify which wasn't
working anyway.

In the small editor I removed strikethrough. The small editor overflowed
the multiple choice edit dialog by a few pixels so I increased its width.