fix passing of shutterbug uri
In production the javascript files are pre-complied. And the SHUTTERBUG_URI is not defined during this stage.
In other parts of LARA we use the 'gon' library to pass variables between ruby and javascript.
So this change switches to that approach.
I haven't run the tests on this, so there might be a following to commit to fix any broken tests.