Use ResizeObserver to monitor interactive's parent size changes
The previous use of ResizeObserver was triggering interactive resizing but
this had no effect when the interactive was pinned because its parent
had a fixed width.

This use of ResizeObserver should really be moved into a runtime library.
It isn't part of authoring and it isn't really part of the interactive-api-parent