This change allows the migration to be re-run even if it failed in the middle
Without this change it would create a second PageItem and Xhtml object for all of the
Pages that had already been processed.

The change uses SQL joins and conditions in order to speed this up.

It also pushes the check for the text being present into the query which will speed it up too.
This check is now just a simple "is not null" which is not the same a present because
present will be false for empty strings. But it still eliminates a lots of pages from the list.

Finally it increases the batch size to 1000. That might help slightly as well since we shouldn't
have memory problems when we are just creating 2 object for each page.
Depending on how find_each is implemented it might be requiring the database to do a full
query each time and just return part of the results. So increasing the batch size should
decrease the total amount of work the database is doing.

[#173020289]