Add Preview In Activity Player link
This adds a link to the top of the activity and page editing pages,
"Preview In Activity Player".

This uses the ACTIVITY_PLAYER_URL environment variable.

The page links to the activity player route introduced in
https://github.com/concord-consortium/activity-player/pull/13
(this route may eventually change)

Note that the `page` url parameter isn't a part of the above PR, but is
added here as future-proofing for when that is added.