Add a default value for PORTAL_PROTOCOL, and error for PORTAL_HOST
PORTAL_PROTOCOL is defined in the environment section of the main docker-compose file,
But that doesn't make it available to the other sections of the file.
Adding a default here prevents SSO errors when it is not defined.

Also this adds a `:?[err]`  to the PORTAL_HOST variable which tells docker it is mandatory