The fix is moving the embeddable_dom_id to the embeddable-root
This fixes the problem because the plugin system uses the element with the embeddable_dom_id
as the output container for the plugin. It also passes the .embeddable-container inside of this
dom element as the wrapped container. Then the plugin adds it content in the output container
and moves the wrapped container div inside of its content. With the previous structure
that meant that the embeddable-root div got lost because it was in between these too divs.

The rest of the changes here are comments to try to explain all of this complexity.