use expect.assertions to make sure async test is completing
I started updating the test to check support in a new unsupported mode.
Doing that, I saw that the test was not actually failing with the unsupported mode.
Instead a different test was failing when the async assertion finally ran and failed.
The expect.assertions keep this from happening.

Additionally moving the mode loop outside of the it block provides more clear
reporting when a specific mode fails.