handle saving managed and mw interactive updating from ITSI form
On the ITSI form there is no page, so there is no page_id parameter
So when changing an interactive in ITSI authoring you see an error message
about "save failed".

This change takes the conservative approach of continuing to use the page_id when it is available.
It seems like this isn't necessary since the other embeddables work like the new code I added.
However perhaps there is a case where update is called before the interactive has been added
to a page.