Convert updated Project form to React
PT Story: https://www.pivotaltracker.com/story/show/181174622

[#181174622]

In addition to converting the Project form to React, these changes also convert the Projects index page. That seemed to make sense since in the current setup, you're redirected there when you create a new project, plus it was easy and I figured it'd be good to convert the whole Project section of the site in one go.