Fix drag-and-drop interactive does not look correct in report/dashboard.
PT Story: https://www.pivotaltracker.com/story/show/181154848

[#181154848]

Adds an optional `view` property for interactives. Currently, this will only be used for determining if an interactive is being displayed as a standalone element in the context of a report or dashboard (i.e., not embedded in the report or dashboard UI).