Added fake glossary plugin to activity export if a glossary model is assigned [#181798706]
This also removes any existing glossary plugin is it is already assigned to the activity.

The first plugin named "glossary" found is used as the values for the fake glossary plugin,
with the s3Url changed to point to the glossary api json endpoint.

This commit also adds a host parameter to the activity and sequence exports as it is needed to resolve the glossary api json url.