Added glossary authoring container [#181873380]
This adds a container div for the glossary authoring along with setting a new GlossaryModelAuthoring namespace in the global LARA variable and then loads the glossary plugin script which will use that info to render into the container.

This also moves the lookup of the glossary approved script into the glossary model so that it can be updated in the future to use a setting instead of just returning the first glossary approved script found.