Added admin settings interface to set current glossary [#181884477]
This adds a generic key/value settings model with a concrete setting controller that currently only sets the approved glossary script id.  It is designed to be extensible to other settings.

To make id mapping more visible the ids have been added to both the plugin list and library interactive list.