Preview interactives in edit form
PT Story: https://www.pivotaltracker.com/story/show/182075235

[#182075235]

These changes modify the assessment item edit form to include a preview of the item being edited. Changes made to the item's settings appear in the preview as they are made. Some settings/features are not yet supported by the preview (these include aspect ratio,  click to play options, and hints).

Also included is a fix for the issue of assessment items on the main page not updating after they are edited until you reload the page. The fix for that is similar to the method used for updating the edit form previews (passing a `resetCount` value down to the iframe component).