Add legacy_ref_type field to managed interactives
PT Story: https://www.pivotaltracker.com/story/show/182786659

[#182786659]

We were originally going to only add a `legacy_ref_id` field which would contain strings that look like `30-MultipleChoice` where "30" is the related built-in embeddable's numeric ID and "MultipleChoice" is the built-in embeddable's type. 

This change will put the two values in separate fields to provide greater flexibility.

Note: The `legacy_ref_id` migration was originally added to the `new-sections` branch of LARA (https://github.com/concord-consortium/lara/pull/984). I copied that migration file to this branch off of `master` since it turns out we are going to need to reference these values before upgrading to LARA 2. So long as the migration file name has the same ID in it (20220711213558) it won't be run again after we merge `new-sections` into `master`. That's why there are two separate migration files in this commit. In the second, we modify the type of `legacy_ref_id` from string to integer, as well as add the `legacy_ref_type` string field.