feat: Add accessibility object to initInterative message [PT-184987026]
This adds an accessibility object to the initInteractive message that currently contains both the name of the required font size (normal or large) and the font size in pixels.

The two font settings allow interactives to either directly set the font size by value or use the font name in their css files.