feat: Added mediaLibrary to initInteractiveMessage [PT-185394064]
This is not used by the old Lara iframe-saver code but is used by AP to gather up media library items and to pass them to interactives like the lab book for use as an image source.

NOTE: there is a chore in the notebook epic to do the final release, this commit uses a pre-release version.