feat: Added font type to accessibility [PT-185336244]
This adds the font type to the accessibility object in the initInteractive message.

This also refactors the useAccessibility hook options to have a top level option that clients can use to have it update the DOM for them.  As part of the new option the client can give a css selector to use to create a font-family rule.  This allows question-interactives to target the inner runtime div.

As with the original accessibility work the LARA scripts that are no longer used have also been updated.