feat: Added project admin UI for user edit page [PT-186125988]
This adds a list of checkboxes, one per project, on the user edit page that can be checked/unchecked to set the projects the user is the admin of.

This also updates the user and project models to have a second project admin has_many association that only holds the user or project.  This allows the standard Rails update_attributes method to update the association based on the checkboxes checked in the form.

Finally, this also updates the styling of the user form, for both the edit and the (unused) new form.